requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
youtube-transcript-api>=1.0.0
yt-dlp>=2023.0.0
